Receiver-SMTP-Server/requirements.txt
aiosmtpd==1.4.2
PyYAML==6.0
mail-parser==3.15.0
